+++ CHANGES.md
@@ -1,3 +1,6 @@
+# 1.9.4
+- Update models to latest
+
 # 1.9.3
 - Add link from config to OpenAI API keys page
 - Update GPT-4 preview model to latest